Solution to Prime Number Checker

+++ Prime Number Checker@@67a4d240c8f64283936a1643/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int isPrime(double a)
+int isPrime(int a)
 {
     return a < 2 ? 0 : (
         ({
